5004: Rename of compaction-job-creation v2 to v1
