+++ java/compaction/compaction-job-creation/src/main/java/sleeper/compaction/job/creationv2/AwsCreateCompactionJobs.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.job.creationv2;
+package sleeper.compaction.job.creation;
 
 import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.sqs.SqsClient;
