+++ java/compaction/compaction-job-creation/src/main/java/sleeper/compaction/job/creationv2/CompactionBatchMessageSenderToSqs.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.job.creationv2;
+package sleeper.compaction.job.creation;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
